#50: Refactor to rd_class.py.
