+++ src/raindrop_todoist_syncer/rd_class.py
@@ -1,0 +1,67 @@
+from datetime import datetime, timezone
+
+
+class Raindrop:
+    """
+    Class for instantiating a Raindrop object from an individual Raindrop JSON.
+
+    Instance variables are those required for:
+
+        a) saving to "tracked_favourites"
+        b) writing to Todoist
+
+    Attributes
+    ----------
+    id : str
+        The unique identifier of the Raindrop.
+    created_time : datetime
+        The time when the Raindrop was created.
+    parsed_time : datetime
+        The time when the Raindrop was parsed by the script.
+    title : str
+        The title of the Raindrop.
+    notes : str
+        Any notes attached to the Raindrop.
+    link : str
+        The hyperlink associated with the Raindrop.
+
+    """
+
+    def __init__(self, raindrop_json: dict) -> None:
+        """
+        # TODO: Add error handling. Raindrop class will crash in event of a missing
+        # TODO  field. (rd_processor has a skipped failing test in the event of a
+        # TODO  raindrop missing a field).
+
+        Instantiate a Raindrop object from an API output JSON for a single Raindrop.
+
+        Parameters
+        ----------
+        raindrop_json : dict
+            The JSON object output by the Raindrop API, representing a single Raindrop.
+        """
+        self.id = raindrop_json["_id"]
+        self.created_time = raindrop_json["created"]
+        self.parsed_time = datetime.now(timezone.utc).isoformat(timespec="milliseconds")
+        self.title = raindrop_json["title"]
+        self.notes = raindrop_json["note"]
+        self.link = raindrop_json["link"]
+
+    def to_dict(self) -> None:
+        """
+        Convert the Raindrop object to a dictionary (for saving to the JSON database).
+
+        Returns
+        -------
+        dict
+            A dictionary representation of the Raindrop object.
+        """
+
+        return {
+            "id": self.id,
+            "created_time": self.created_time,
+            "parsed_time": self.parsed_time,
+            "title": self.title,
+            "notes": self.notes,
+            "link": self.link,
+        }
