+++ tests/conftest.py
@@ -8,11 +8,11 @@ import pytest
 from requests import HTTPError
 
 from raindrop_todoist_syncer.raindrop import (
-    Raindrop,
     RaindropAccessTokenRefresher,
     RaindropCredentialsManager,
     EnvironmentVariablesFileManager,
 )
+from raindrop_todoist_syncer.rd_class import Raindrop
 
 # ---------------------------------- env vars-------------------------------------------
